Lecture 48 - working with "state".
Component functions are called once during the initial rendering.
To tell React to run the component function again to update some state,
call React hook functions from inside of component functions:
    import React, { useState } from "react";

    function MyComponent(props) {
        const [variable, setVariable] = useState(initial value);
        const clickHandler = () => {
            setVariable(new value);
        };
        return (
            <div>
                <h1 onClick={clickHandler}>
                    {title}
                </h1>
            </div>
        );
    }

- title will have the initial value that was passed to useState()
- React hook functions start with "use"
- calling setVariable(new value) does not change the value right away,
but schedules the state update

Use the useState() hook whenever there is data that might change and
changes to that data need to be reflected in the user interface.