Lecture 49 - a closer look at the "useState" hook.
Added console log to show that component functions are called once
when initially rendered and that state updates tell React to call
them again.

Each instance of a component has it's own state. Setting the state of
one instance of, i.e. MyComponent, does not change the state of other
instances of MyComponent.