Lecture 53 - working with multiple states.
React states can be used to keep track of data that should persist
even after the component function gets called again.

Components can keep track of multiple state variables by calling
useState() for each variable it wants to keep independent of the
component's life cycle.

    function MyComponent(props) {
        const [var1, setVar1] = useState(some default value);
        const [var2, setVar2] = useState(some other default value);
        ...
    }

Note: the convention for using useState() is to name the returned
function with the word "set" as a prefix to the variable name as in:
    const [myState, setMyState] = useState('');