Lecture 54 - using one state instead (and what's better)
An alternative to calling useState() multiple times is to call it
once and pass in an object as the initial value as such:
    const [values, setValues] = useState({
        var1: val1,
        var2: val2
    });

If this method is used, remember to also include the values of ALL the
properies in the object when setting any one specific value. Otherwise,
React will replace the old state object with a new object that
only contains the specific property that was set. The old object keys
would be lost.

Example of setting var1 in the state above:
    setValues({
        var1: newValue1,
        var2: values.var2
    });

    OR

    setValues({
        ...values,
        var1, newValue1
    });

- without explicitly including the val2 key and copying it's value into
the object passed to setValues(), the values state would only contain:
    { var1: newValue1 }