Lecture 55 - updating state that depends on the previous state.
When scheduling lots of state updates or when one state depends on
another state, it is possible that the current state snapshot could be
outdated or incorrect. In this case, it is better to pass a function
to the state's setter function that will be called with the previous
state passed in as an argument.